Abstract detail page factory
