+++ network-api/networkapi/wagtailpages/factory/libraries/research_hub/detail_page.py
@@ -1,41 +1,23 @@
 import random
 
 import factory
-import wagtail_factories
 
-from networkapi.utility.faker import helpers as faker_helpers
 from networkapi.wagtailpages import models as wagtailpage_models
-from networkapi.wagtailpages.factory import documents as documents_factory
-from networkapi.wagtailpages.factory import image_factory
+from networkapi.wagtailpages.factory.libraries import (
+    detail_page as library_detail_page_factories,
+)
 
 
-class ResearchDetailPageFactory(wagtail_factories.PageFactory):
+class ResearchDetailPageFactory(library_detail_page_factories.LibraryDetailPageAbstractFactory):
     class Meta:
         model = wagtailpage_models.ResearchDetailPage
 
-    title = factory.Faker("text", max_nb_chars=50)
-    cover_image = factory.SubFactory(image_factory.ImageFactory)
     links = factory.RelatedFactoryList(
         factory="networkapi.wagtailpages.factory.libraries.research_hub.detail_page.ResearchDetailLinkFactory",
         factory_related_name="detail_page",
         size=lambda: random.randint(1, 2),
         with_url=True,
     )
-    original_publication_date = factory.Faker("date_object")
-    introduction = factory.Faker("text", max_nb_chars=300)
-
-    @factory.lazy_attribute
-    def overview(self):
-        faker = faker_helpers.get_faker()
-        return "\n\n".join(faker.paragraphs(nb=3))
-
-    @factory.lazy_attribute
-    def collaborators(self):
-        faker = faker_helpers.get_faker()
-        names = []
-        for _ in range(random.randint(1, 5)):
-            names.append(faker.name())
-        return "; ".join(names)
 
     authors = factory.RelatedFactoryList(
         factory="networkapi.wagtailpages.factory.libraries.research_hub.relations.ResearchAuthorRelationFactory",
@@ -62,18 +44,6 @@ class ResearchDetailPageFactory(wagtail_factories.PageFactory):
     )
 
 
-class ResearchDetailLinkFactory(factory.django.DjangoModelFactory):
+class ResearchDetailLinkFactory(library_detail_page_factories.LibraryDetailLinkBaseAbstractFactory):
     class Meta:
         model = wagtailpage_models.ResearchDetailLink
-
-    label = factory.Faker("text", max_nb_chars=30)
-    url = ""
-    page = None
-    document = None
-
-    class Params:
-        with_url = factory.Trait(
-            url=factory.Faker("uri"),
-        )
-        with_page = factory.Trait(page=factory.SubFactory(wagtail_factories.PageFactory))
-        with_document = factory.Trait(document=factory.SubFactory(documents_factory.DocumentFactory))
